pyproject.toml
[tool.poetry]
name = "django-saas-launcher"
version = "1.0.0"
description = "Django SaaS Launcher is a tool that is designed to launch a SaaS system."
license = "BSD-3-Clause"
authors = ["LaunchAt, Inc. <hello@launchat.jp>"]
maintainers = []
readme = "README.md"
homepage = "https://launchat.github.io/django-saas-launcher"
repository = "https://github.com/LaunchAt/django-saas-launcher"
documentation = "https://launchat.github.io/django-saas-launcher"
keywords = []
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{ include = "launcher" }]
include = []
exclude = []

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-hosts = "^5.2"
django-storages = "^1.13.2"
boto3 = "^1.26.71"
watchtower = "^3.0.1"
phonenumbers = "^8.13.6"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.0.12"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.1.0"
mypy = "^1.0.0"
pytest = "^7.2.1"
tox = "^4.4.5"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.25.1"
django-filter = "^22.1"

[build-system]
requires = ["poetry-core>=1.2.2"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310", "py311"]
skip-string-normalization = true
extend-exclude = "^.*/migrations/*$"

[tool.isort]
extend_skip = [".tox"]
extend_skip_glob = ["*/migrations/*"]
skip_gitignore = true
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
exclude = ["^.*/migrations/*$"]

[tool.pytest.ini_options]
testpaths = ["tests"]
